+++ projects/qgrid-core/infrastructure/defer.d.ts
@@ -1,11 +1,11 @@
-export declare class DeferPromise {
-	catch(handler);
-	then(handler);
+export declare class DeferPromise<T> {
+	catch(handler: () => void): DeferPromise<T>;
+	then(handler: (value: T) => void): DeferPromise<T>;
 }
 
-export declare class Defer {
-	promise: DeferPromise;
-	reject();
-	resolve();
+export declare class Defer<T> {
+	promise: DeferPromise<T>;
+	reject(): void;
+	resolve(value: T): void;
 }