+++ projects/qgrid-core/services/dom.d.ts
@@ -1,4 +1,4 @@
 export declare function css(element: HTMLElement, property: string, value: string): string;
 export declare function elementFromPoint(x: number, y: number): HTMLElement;
-export declare function eventPath(event: MouseEvent): HTMLElement[];
+export declare function eventPath(event: MouseEvent | KeyboardEvent): HTMLElement[];
 export declare function parents(element: HTMLElement): HTMLElement[];