+++ projects/qgrid-core/services/job.line.d.ts
@@ -3,5 +3,5 @@
 export declare function jobLine(delay: number): (job: () => void) => Promise<void>;
 
 export namespace jobLine {
-	function run(job: any, delay: number): Defer;
+	function run(job: any, delay: number): Defer<any>;
 }